Make timestamps update live
Also: factor out relative date display into a separate component
