+++ client/src/components/feeds/post/Post.tsx
@@ -1,6 +1,6 @@
 import "./Post.css";
 import ContentWarning from "./ContentWarning";
-import PostDate from "./content/PostDate";
+import RelativeDate from "./RelativeDate";
 import PostNumber from "./content/PostNumber";
 import ReactionBar from "./reactions/ReactionBar";
 import DividerDot from "./content/DividerDot";
@@ -78,15 +78,15 @@ function Post(props: PostProps) {
             <ContentWarning ContentWarningText={props.post.contentWarning} />
           )}
         </div>
-        <PostDate
-          value={
-            new Date(
+        <div className="PostDate">
+          <RelativeDate
+            date={
               props.post.needsReview
                 ? props.post.postTime
                 : props.post.approvedTime
-            )
-          }
-        />
+            }
+          />
+        </div>
       </div>
       <div className="PostBody">
         <UserContent showContent={!props.post.contentWarning}>
