+++ client/src/components/feeds/post/RelativeDate.css
@@ -1,0 +1,5 @@
+.RelativeDate {
+  color: #888;
+  user-select: none;
+  cursor: default;
+}
