+++ client/src/components/feeds/post/comments/comment_header/CommentHeader.css
@@ -19,10 +19,7 @@
 }
 
 .CommentDate {
-  color: gray;
   font-size: 0.8rem;
-  user-select: none;
-  cursor: default;
 }
 
 .ExpandThreadButton {
@@ -38,7 +35,7 @@
 
 .DevBadge {
   margin-left: -0.6rem;
-  font-size: .7rem;
+  font-size: 0.7rem;
   background-color: #62c6e6;
   color: white;
   border-radius: 2px;
