--- client/src/components/feeds/post/content/PostDate.css
@@ -1,7 +1,0 @@
-.PostDate {
-  margin: 0 0 0.6em 0;
-  color: #999;
-  float: right;
-  user-select: none;
-  cursor: default;
-}
